requirements.txt
Pillow==9.5.0
pillow-heif==0.10.0
piexif==1.1.3
imagehash==4.3.1
Wand==0.6.11
opencv-python==4.7.0.72
numpy==1.24.3
tqdm==4.65.0